requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
black==25.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dataclasses-json==0.6.7
distlib==0.4.0
distro==1.9.0
filelock==3.25.2
frozenlist==1.8.0
greenlet==3.3.2
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain== *
langchain-community==0.3.31
langchain-core== *
langchain-groq==0.3.8
langchain-openai==0.3.35
langchain-tavily==0.2.17
langchain-text-splitters==0.3.11
langgraph==1.1.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.18
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
openai==2.28.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pathspec==1.0.4
pipenv==2026.1.0
platformdirs==4.9.4
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyjokes==0.8.3
python-discovery==1.1.3
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==82.0.1
sniffio==1.3.1
SQLAlchemy==2.0.48
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
virtualenv==21.2.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
